+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/dto/expenditure/CreateExpenditureRequest.java
@@ -1,19 +1,17 @@
-package com.prgrms.tenwonmoa.domain.accountbook.dto;
+package com.prgrms.tenwonmoa.domain.accountbook.dto.expenditure;
 
 import java.time.LocalDateTime;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
 import com.prgrms.tenwonmoa.domain.accountbook.Expenditure;
 import com.prgrms.tenwonmoa.domain.category.UserCategory;
 import com.prgrms.tenwonmoa.domain.user.User;
 
-import lombok.Getter;
-
-@Getter
-public class UpdateExpenditureRequest {
+public class CreateExpenditureRequest {
 
 	@NotNull
 	private final LocalDateTime registerDate;
@@ -22,12 +20,13 @@ public class UpdateExpenditureRequest {
 	@Max(1000000000000L)
 	private final Long amount;
 
+	@Size(max = 50)
 	private final String content;
 
 	@NotNull
 	private final Long userCategoryId;
 
-	public UpdateExpenditureRequest(LocalDateTime registerDate, Long amount, String content, Long userCategoryId) {
+	public CreateExpenditureRequest(LocalDateTime registerDate, Long amount, String content, Long userCategoryId) {
 		this.registerDate = registerDate;
 		this.amount = amount;
 		this.content = content;
