+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/dto/expenditure/FindExpenditureResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.tenwonmoa.domain.accountbook.dto;
+package com.prgrms.tenwonmoa.domain.accountbook.dto.expenditure;
 
 import java.time.LocalDateTime;
