+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/dto/expenditure/UpdateExpenditureRequest.java
@@ -1,42 +1,41 @@
-package com.prgrms.tenwonmoa.domain.accountbook.dto;
+package com.prgrms.tenwonmoa.domain.accountbook.dto.expenditure;
 
 import java.time.LocalDateTime;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
 
-import com.prgrms.tenwonmoa.domain.accountbook.Income;
+import com.prgrms.tenwonmoa.domain.accountbook.Expenditure;
 import com.prgrms.tenwonmoa.domain.category.UserCategory;
 import com.prgrms.tenwonmoa.domain.user.User;
 
 import lombok.Getter;
 
 @Getter
-public class UpdateIncomeRequest {
+public class UpdateExpenditureRequest {
+
 	@NotNull
 	private final LocalDateTime registerDate;
 
-	@Min(1L)
-	@Max(1_000_000_000_000L)
+	@Min(0L)
+	@Max(1000000000000L)
 	private final Long amount;
 
-	@Size(max = 50)
 	private final String content;
 
 	@NotNull
 	private final Long userCategoryId;
 
-	public UpdateIncomeRequest(LocalDateTime registerDate, Long amount, String content, Long userCategoryId) {
+	public UpdateExpenditureRequest(LocalDateTime registerDate, Long amount, String content, Long userCategoryId) {
 		this.registerDate = registerDate;
 		this.amount = amount;
 		this.content = content;
 		this.userCategoryId = userCategoryId;
 	}
 
-	public Income toEntity(User user, UserCategory userCategory, String categoryName) {
-		return new Income(
+	public Expenditure toEntity(User user, UserCategory userCategory, String categoryName) {
+		return new Expenditure(
 			this.registerDate,
 			this.amount,
 			this.content,
@@ -44,5 +43,9 @@ public class UpdateIncomeRequest {
 			user,
 			userCategory
 		);
+	}
+
+	public Long getUserCategoryId() {
+		return userCategoryId;
 	}
 }
