+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/dto/income/CreateIncomeRequest.java
@@ -1,4 +1,4 @@
-package com.prgrms.tenwonmoa.domain.accountbook.dto;
+package com.prgrms.tenwonmoa.domain.accountbook.dto.income;
 
 import java.time.LocalDateTime;
