+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/dto/income/FindIncomeResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.tenwonmoa.domain.accountbook.dto;
+package com.prgrms.tenwonmoa.domain.accountbook.dto.income;
 
 import java.time.LocalDateTime;
