+++ src/main/java/com/prgrms/tenwonmoa/domain/accountbook/service/ExpenditureService.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.prgrms.tenwonmoa.domain.accountbook.Expenditure;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.CreateExpenditureRequest;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.CreateExpenditureResponse;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.FindExpenditureResponse;
-import com.prgrms.tenwonmoa.domain.accountbook.dto.UpdateExpenditureRequest;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.expenditure.CreateExpenditureRequest;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.expenditure.CreateExpenditureResponse;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.expenditure.FindExpenditureResponse;
+import com.prgrms.tenwonmoa.domain.accountbook.dto.expenditure.UpdateExpenditureRequest;
 import com.prgrms.tenwonmoa.domain.accountbook.repository.ExpenditureRepository;
 import com.prgrms.tenwonmoa.domain.category.Category;
 import com.prgrms.tenwonmoa.domain.category.UserCategory;
